+++ meal_db.py
@@ -16,14 +16,16 @@ def create_table():
     )""")
 
 def add_meal(food):
-    with conn:
-        c.execute("INSERT INTO menu (meal, timestamp) VALUES (%s, %s)", (food, time.time()))
-    sys.stdout.write('Added %s to DB' % food)
+    if food:
+        with conn:
+            c.execute("INSERT INTO menu (meal, timestamp) VALUES (%s, %s)", (food, time.time()))
+        sys.stdout.write('Added %s to DB' % food)
 
 def remove_meal(food):
-    with conn:
-        c.execute("DELETE FROM menu WHERE meal = %s", (food,))
-    sys.stdout.write('Removed %s from DB' % food)
+    if food:
+        with conn:
+            c.execute("DELETE FROM menu WHERE meal = %s", (food,))
+        sys.stdout.write('Removed %s from DB' % food)
 
 def return_list_of_all_meals():
     c.execute("SELECT meal FROM menu ORDER BY timestamp DESC")
